array[]为uint8_t时sizeof(array)=4,为16_t时也为4，但单元素从1变2

+++ Hardware/Serial.h
@@ -3,5 +3,5 @@
 void Serial_Initilize(void);
 void Serial_SendByte(uint8_t data);
 void Serial_SendString(uint8_t *str);
-void Serial_SendArray(uint8_t *array);
+void Serial_SendArray(uint16_t *array);
 #endif
